Refactor the calypso-e2e library.
- add docstrings where it was missing.
- move some functions out of browser-manager to browser-helper (new file).
- extract filename building process in screenshots to a function in media-helper.
- add user agent string to the browser launch process.
