+++ packages/calypso-e2e/src/browser-manager.ts
@@ -1,5 +1,5 @@
 /**
- * @file Manages instance of Playwright Browser and BrowserContext.
+ * @file Helps manage browser instance.
  * @author Edwin Takahashi
  */
 
@@ -14,62 +14,14 @@ import type { Browser, BrowserContext, Page } from 'playwright';
  * Internal dependencies
  */
 import { getVideoDir } from './media-helper';
-import type { screenSize, localeCode } from './types';
+import { getScreenDimension } from './browser-helper';
 
+/**
+ * Constants
+ */
 const playwrightTimeoutMS: number = config.get( 'playwrightTimeoutMS' );
 
 export let browser: Browser;
-
-/**
- * Returns the target screen size for tests to run against.
- *
- * By default, this function will return 'desktop' as the target.
- * To specify another screen size, set the BROWSERSIZE environment variable.
- *
- * @returns {screenSize} Target screen size.
- */
-export function getTargetScreenSize(): screenSize {
-	return ! process.env.BROWSERSIZE
-		? 'desktop'
-		: ( process.env.BROWSERSIZE.toLowerCase() as screenSize );
-}
-
-/**
- * Returns the locale under test.
- *
- * By default, this function will return 'en' as the locale.
- * To set the locale, set the BROWSERLOCALE environment variable.
- *
- * @returns {localeCode} Target locale code.
- */
-export function getTargetLocale(): localeCode {
-	return ! process.env.BROWSERLOCALE ? 'en' : process.env.BROWSERLOCALE.toLowerCase();
-}
-
-/**
- * Returns a set of screen dimensions in numbers.
- *
- * This function takes the output of `getTargetScreenSize` and returns an
- * object key/value mapping of the screen diemensions represented by
- * the output.
- *
- * @returns {number, number} Object with key/value mapping of screen dimensions.
- * @throws {Error} If target screen size was not set.
- */
-export function getScreenDimension(): { width: number; height: number } {
-	switch ( getTargetScreenSize() ) {
-		case 'mobile':
-			return { width: 400, height: 1000 };
-		case 'tablet':
-			return { width: 1024, height: 1000 };
-		case 'desktop':
-			return { width: 1440, height: 1000 };
-		case 'laptop':
-			return { width: 1400, height: 790 };
-		default:
-			throw new Error( 'Unsupported screen size specified.' );
-	}
-}
 
 /**
  * Familiar entrypoint to initialize the browser from a test writer's perspective.
@@ -114,11 +66,13 @@ export async function launchBrowserContext(): Promise< BrowserContext > {
 	// By default, record video for each browser context.
 	const videoDir = getVideoDir();
 	const dimension = getScreenDimension();
+	const userAgent = `user-agent=Mozilla/5.0 (wp-e2e-tests) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/${ await browser.version() } Safari/537.36`;
 
 	// Generate a new BrowserContext.
 	return await browser.newContext( {
 		viewport: null, // Do not override window size set in the browser launch parameters.
 		recordVideo: { dir: videoDir, size: { width: dimension.width, height: dimension.height } },
+		userAgent: userAgent,
 	} );
 }
