+++ packages/calypso-e2e/src/index.ts
@@ -3,4 +3,5 @@
  */
 export * as BrowserManager from './browser-manager';
 export * as MediaHelper from './media-helper';
+export * as BrowserHelper from './browser-helper';
 export * from './hooks';
